can all you fucking kids stop moaning now for fuck sake
no its not working with fucking Panorama as i couldnt really care and dont paste anymore
use this shit if you want to use Panorama no more updates
https://pasters.cc/showthread.php?tid=5692

+++ INDIGO/Engine/Weapon.h
@@ -30,7 +30,7 @@ namespace Engine
 	{
 	public:
 //[swap_lines]
-		int*			GetItemDefinitionIndex();
+		short*			GetItemDefinitionIndex();
 		int*			GetItemIDHigh();
 		int*			GetAccountID();
 		int*			GetEntityQuality();
